+++ library/src/com/astuetz/PagerSlidingTabStrip.java
@@ -22,6 +22,7 @@ import android.content.res.ColorStateList;
 import android.content.res.TypedArray;
 import android.database.DataSetObserver;
 import android.graphics.Canvas;
+import android.graphics.Color;
 import android.graphics.Paint;
 import android.graphics.Paint.Style;
 import android.graphics.Typeface;
@@ -29,7 +30,6 @@ import android.os.Build;
 import android.os.Parcel;
 import android.os.Parcelable;
 import android.support.v4.util.Pair;
-import android.support.v4.view.ViewCompat;
 import android.support.v4.view.ViewPager;
 import android.support.v4.view.ViewPager.OnPageChangeListener;
 import android.util.AttributeSet;
@@ -48,9 +48,6 @@ import com.astuetz.pagerslidingtabstrip.R;
 import java.util.Locale;
 
 public class PagerSlidingTabStrip extends HorizontalScrollView {
-
-    private static final float OPAQUE = 1.0f;
-    private static final float HALF_TRANSP = 0.5f;
 
     public interface CustomTabProvider {
         public View getCustomTabView(ViewGroup parent, int position);
@@ -112,8 +109,8 @@ public class PagerSlidingTabStrip extends HorizontalScrollView {
     private int tabPadding = 12;
     private int tabTextSize = 14;
     private ColorStateList tabTextColor = null;
-    private float tabTextAlpha = HALF_TRANSP;
-    private float tabTextSelectedAlpha = OPAQUE;
+    private ColorStateList tabTextColorSelected = null;
+    private int textAlpha = 150;
 
     private int paddingLeft = 0;
     private int paddingRight = 0;
@@ -164,11 +161,6 @@ public class PagerSlidingTabStrip extends HorizontalScrollView {
         tabTextSize = a.getDimensionPixelSize(TEXT_SIZE_INDEX, tabTextSize);
         ColorStateList colorStateList = a.getColorStateList(TEXT_COLOR_INDEX);
         int textPrimaryColor = a.getColor(TEXT_COLOR_PRIMARY, android.R.color.white);
-        if (colorStateList != null) {
-            tabTextColor = colorStateList;
-        } else {
-            tabTextColor = getColorStateList(textPrimaryColor);
-        }
 
         underlineColor = textPrimaryColor;
         dividerColor = textPrimaryColor;
@@ -195,16 +187,22 @@ public class PagerSlidingTabStrip extends HorizontalScrollView {
         isPaddingMiddle = a.getBoolean(R.styleable.PagerSlidingTabStrip_pstsPaddingMiddle, isPaddingMiddle);
         tabTypefaceStyle = a.getInt(R.styleable.PagerSlidingTabStrip_pstsTextStyle, Typeface.BOLD);
         tabTypefaceSelectedStyle = a.getInt(R.styleable.PagerSlidingTabStrip_pstsTextSelectedStyle, Typeface.BOLD);
-        tabTextAlpha = a.getFloat(R.styleable.PagerSlidingTabStrip_pstsTextAlpha, HALF_TRANSP);
-        tabTextSelectedAlpha = a.getFloat(R.styleable.PagerSlidingTabStrip_pstsTextSelectedAlpha, OPAQUE);
+        tabTextColorSelected = a.getColorStateList(R.styleable.PagerSlidingTabStrip_pstsTextColorSelected);
+        textAlpha = a.getInt(R.styleable.PagerSlidingTabStrip_pstsTextAlpha, textAlpha);
         a.recycle();
+
+        tabTextColor = colorStateList == null ? getColorStateList(Color.argb(textAlpha,
+                Color.red(textPrimaryColor),
+                Color.green(textPrimaryColor),
+                Color.blue(textPrimaryColor))) : colorStateList;
+
+        tabTextColorSelected = tabTextColorSelected == null ? getColorStateList(textPrimaryColor) : tabTextColorSelected;
 
         setMarginBottomTabContainer();
 
         rectPaint = new Paint();
         rectPaint.setAntiAlias(true);
         rectPaint.setStyle(Style.FILL);
-
 
         dividerPaint = new Paint();
         dividerPaint.setAntiAlias(true);
@@ -280,8 +278,8 @@ public class PagerSlidingTabStrip extends HorizontalScrollView {
         TextView textView = (TextView) tabView.findViewById(R.id.tab_title);
         if (textView != null) {
             if (title != null) textView.setText(title);
-            float alpha = pager.getCurrentItem() == position ? tabTextSelectedAlpha : tabTextAlpha;
-            ViewCompat.setAlpha(textView, alpha);
+            ColorStateList color = pager.getCurrentItem() == position ? tabTextColorSelected : tabTextColor;
+            textView.setTextColor(color);
         }
 
         tabView.setFocusable(true);
@@ -461,7 +459,7 @@ public class PagerSlidingTabStrip extends HorizontalScrollView {
             if (state == ViewPager.SCROLL_STATE_IDLE) {
                 scrollToChild(pager.getCurrentItem(), 0);
             }
-            //Full alpha for current item
+            //Full textAlpha for current item
             View currentTab = tabsContainer.getChildAt(pager.getCurrentItem());
             selected(currentTab);
             //Half transparent for prev item
@@ -508,7 +506,7 @@ public class PagerSlidingTabStrip extends HorizontalScrollView {
             TextView title = (TextView) tab.findViewById(R.id.tab_title);
             if (title != null) {
                 title.setTypeface(tabTypeface, tabTypefaceStyle);
-                ViewCompat.setAlpha(title, tabTextAlpha);
+                title.setTextColor(tabTextColor);
             }
         }
     }
@@ -518,7 +516,7 @@ public class PagerSlidingTabStrip extends HorizontalScrollView {
             TextView title = (TextView) tab.findViewById(R.id.tab_title);
             if (title != null) {
                 title.setTypeface(tabTypeface, tabTypefaceSelectedStyle);
-                ViewCompat.setAlpha(title, tabTextSelectedAlpha);
+                title.setTextColor(tabTextColorSelected);
             }
         }
     }