+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Change Log
 ==========
 Version 1.0.9 *(-- - -- - ----)*
 ---------------------------------
+* Bug fixes
 * Add parsing of independent padding attributes (#69)
 * pstsTextAlpha value range change from 0..1 to 0..255.
 * New attr pstsTextColorSelected